# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BusTrack API — Rural Bus Booking & Live Tracking
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Stop {
  id: number;
  name: string;
  order: number;
  estimatedMinutes: number;
  distanceKm?: number;
}

export type BusStatus = typeof BusStatus[keyof typeof BusStatus];


export const BusStatus = {
  scheduled: 'scheduled',
  running: 'running',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Bus {
  id: number;
  busNumber: string;
  routeName: string;
  from: string;
  to: string;
  departureTime: string;
  arrivalTime?: string;
  totalSeats: number;
  availableSeats: number;
  fare: number;
  status: BusStatus;
}

export type BusDetailStatus = typeof BusDetailStatus[keyof typeof BusDetailStatus];


export const BusDetailStatus = {
  scheduled: 'scheduled',
  running: 'running',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface BusDetail {
  id: number;
  busNumber: string;
  routeName: string;
  from: string;
  to: string;
  departureTime: string;
  arrivalTime?: string;
  totalSeats: number;
  availableSeats: number;
  fare: number;
  status: BusDetailStatus;
  stops: Stop[];
}

export type SeatType = typeof SeatType[keyof typeof SeatType];


export const SeatType = {
  window: 'window',
  aisle: 'aisle',
  middle: 'middle',
} as const;

export type SeatStatus = typeof SeatStatus[keyof typeof SeatStatus];


export const SeatStatus = {
  available: 'available',
  booked: 'booked',
  reserved: 'reserved',
} as const;

export interface Seat {
  id: number;
  seatNumber: string;
  row: number;
  col: number;
  type: SeatType;
  status: SeatStatus;
}

export type TrackingInfoStatus = typeof TrackingInfoStatus[keyof typeof TrackingInfoStatus];


export const TrackingInfoStatus = {
  scheduled: 'scheduled',
  running: 'running',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

/**
 * @nullable
 */
export type TrackingInfoNextStop = {
  id?: number;
  name?: string;
  order?: number;
  estimatedMinutes?: number;
  distanceKm?: number;
} | null;

export interface TrackingInfo {
  busId: number;
  busNumber: string;
  routeName: string;
  from: string;
  to: string;
  status: TrackingInfoStatus;
  currentStop: Stop;
  /** @nullable */
  nextStop: TrackingInfoNextStop;
  /** @nullable */
  etaMinutes?: number | null;
  progressPercent: number;
  lastUpdated: string;
}

export interface BusSummary {
  totalBuses: number;
  runningNow: number;
  scheduledToday: number;
  totalBookingsToday: number;
  availableSeatsTotal: number;
}

export interface BookingInput {
  busId: number;
  seatId: number;
  passengerName: string;
  phone: string;
  journeyDate: string;
  /** @nullable */
  boardingStop?: string | null;
  /** @nullable */
  dropStop?: string | null;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  busId: number;
  seatId: number;
  passengerName: string;
  phone: string;
  journeyDate: string;
  /** @nullable */
  boardingStop?: string | null;
  /** @nullable */
  dropStop?: string | null;
  status: BookingStatus;
  createdAt: string;
  bus: Bus;
  seat: Seat;
}

export type ListBusesParams = {
from?: string;
to?: string;
date?: string;
};