# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BusTrack API — Rural Bus Booking & Live Tracking
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';
import type { Bus } from './bus';
import type { Seat } from './seat';

export interface Booking {
  id: number;
  busId: number;
  seatId: number;
  passengerName: string;
  phone: string;
  journeyDate: string;
  /** @nullable */
  boardingStop?: string | null;
  /** @nullable */
  dropStop?: string | null;
  status: BookingStatus;
  createdAt: string;
  bus: Bus;
  seat: Seat;
}