# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BusTrack API — Rural Bus Booking & Live Tracking
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  busId: number;
  seatId: number;
  passengerName: string;
  phone: string;
  journeyDate: string;
  /** @nullable */
  boardingStop?: string | null;
  /** @nullable */
  dropStop?: string | null;
}