# lib/api-zod/src/generated/types/bus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BusTrack API — Rural Bus Booking & Live Tracking
 * OpenAPI spec version: 0.1.0
 */
import type { BusStatus } from './busStatus';

export interface Bus {
  id: number;
  busNumber: string;
  routeName: string;
  from: string;
  to: string;
  departureTime: string;
  arrivalTime?: string;
  totalSeats: number;
  availableSeats: number;
  fare: number;
  status: BusStatus;
}