# lib/api-zod/src/generated/types/busDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BusTrack API — Rural Bus Booking & Live Tracking
 * OpenAPI spec version: 0.1.0
 */
import type { BusDetailStatus } from './busDetailStatus';
import type { Stop } from './stop';

export interface BusDetail {
  id: number;
  busNumber: string;
  routeName: string;
  from: string;
  to: string;
  departureTime: string;
  arrivalTime?: string;
  totalSeats: number;
  availableSeats: number;
  fare: number;
  status: BusDetailStatus;
  stops: Stop[];
}