# lib/api-zod/src/generated/types/seat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BusTrack API — Rural Bus Booking & Live Tracking
 * OpenAPI spec version: 0.1.0
 */
import type { SeatStatus } from './seatStatus';
import type { SeatType } from './seatType';

export interface Seat {
  id: number;
  seatNumber: string;
  row: number;
  col: number;
  type: SeatType;
  status: SeatStatus;
}