# lib/api-zod/src/generated/types/stop.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BusTrack API — Rural Bus Booking & Live Tracking
 * OpenAPI spec version: 0.1.0
 */

export interface Stop {
  id: number;
  name: string;
  order: number;
  estimatedMinutes: number;
  distanceKm?: number;
}