# lib/api-zod/src/generated/types/trackingInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BusTrack API — Rural Bus Booking & Live Tracking
 * OpenAPI spec version: 0.1.0
 */
import type { Stop } from './stop';
import type { TrackingInfoNextStop } from './trackingInfoNextStop';
import type { TrackingInfoStatus } from './trackingInfoStatus';

export interface TrackingInfo {
  busId: number;
  busNumber: string;
  routeName: string;
  from: string;
  to: string;
  status: TrackingInfoStatus;
  currentStop: Stop;
  /** @nullable */
  nextStop: TrackingInfoNextStop;
  /** @nullable */
  etaMinutes?: number | null;
  progressPercent: number;
  lastUpdated: string;
}